feat(satp-hermes): update constants
runs linter as well
