+++ packages/cactus-plugin-satp-hermes/src/main/typescript/core/constants.ts
@@ -42,18 +42,13 @@ export const DEFAULT_PORT_GATEWAY_UI = DEFAULT_PORT_GATEWAY_SERVER + 2;
 /** Default port for SATP gateway OpenAPI documentation */
 export const DEFAULT_PORT_GATEWAY_OAPI = 4010;
 /** Current SATP protocol version identifier */
-export const SATP_VERSION = "v13";
+export const SATP_VERSION = "v13.0.1";
 /** SATP Core specification version */
 export const SATP_CORE_VERSION = "v13";
 /** SATP Architecture specification version */
-export const SATP_ARCHITECTURE_VERSION = "v13";
+export const SATP_ARCHITECTURE_VERSION = "v09";
 /** SATP Crash Recovery specification version */
-export const SATP_CRASH_VERSION = "v13";
-/**
- * SATP protocol wire version string per v13 Section 5.3.1.
- * Carried in the `version` field of CommonSatp.
- */
-export const SATP_PROTOCOL_VERSION = "1.0";
+export const SATP_CRASH_VERSION = "v06";
 /** Mandatory-to-implement TLS 1.3 cipher suite from RFC 8446 Section 9.1. */
 export const DEFAULT_TLS13_CIPHER_SUITE = "TLS_AES_128_GCM_SHA256";
