+++ packages/cactus-plugin-satp-hermes/src/test/typescript/constants.ts
@@ -1,3 +1,3 @@
-// TODO replace with the official after the release
-export const SATP_DOCKER_IMAGE_NAME = "rafaelapb/cacti-satp-hermes-gateway";
-export const SATP_DOCKER_IMAGE_VERSION = "2026-06-17-dev-01760ca80";
+export const SATP_DOCKER_IMAGE_NAME =
+  "hyperledger-cacti/cacti-satp-hermes-gateway";
+export const SATP_DOCKER_IMAGE_VERSION = "3.0.1-dev.d68c9ae";
