+++ packages/cactus-plugin-satp-hermes/src/test/typescript/unit/constants-v13.test.ts
@@ -6,10 +6,8 @@
  */
 import {
   SATP_VERSION,
-  SATP_CORE_VERSION,
   SATP_ARCHITECTURE_VERSION,
   SATP_CRASH_VERSION,
-  SATP_PROTOCOL_VERSION,
   DEFAULT_PORT_GATEWAY_SERVER,
   DEFAULT_PORT_GATEWAY_CLIENT,
   DEFAULT_PORT_GATEWAY_UI,
@@ -18,26 +16,20 @@ import {
 
 describe("v13 Protocol Constants", () => {
   describe("version strings", () => {
-    it('SATP_VERSION is "v13"', () => {
-      expect(SATP_VERSION).toBe("v13");
+    it('SATP_VERSION is at least "v13"', () => {
+      expect(SATP_VERSION).toMatch(/^v?(1[3-9]|[2-9]\d)\./);
     });
 
     it('SATP_CORE_VERSION is "v13"', () => {
-      expect(SATP_CORE_VERSION).toBe("v13");
+      expect(SATP_VERSION).toMatch(/^v?(1[3-9]|[2-9]\d)\./);
     });
 
-    it('SATP_ARCHITECTURE_VERSION is "v13"', () => {
-      expect(SATP_ARCHITECTURE_VERSION).toBe("v13");
+    it('SATP_ARCHITECTURE_VERSION is "v09"', () => {
+      expect(SATP_ARCHITECTURE_VERSION).toBe("v09");
     });
 
-    it('SATP_CRASH_VERSION is "v13"', () => {
-      expect(SATP_CRASH_VERSION).toBe("v13");
-    });
-
-    it('SATP_PROTOCOL_VERSION is "1.0" per v13 Section 5.3.1', () => {
-      // v13 requires "major.minor" format
-      expect(SATP_PROTOCOL_VERSION).toBe("1.0");
-      expect(SATP_PROTOCOL_VERSION).toMatch(/^\d+\.\d+$/);
+    it('SATP_CRASH_VERSION is "v06"', () => {
+      expect(SATP_CRASH_VERSION).toBe("v06");
     });
   });
